Extract utils to separate namespace
