+++ init/utils.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const SIZE_UNITS = ['', ' Kb', ' Mb', ' Gb', ' Tb', ' Pb', ' Eb', ' Zb', ' Yb'];
+
+const bytesToSize = bytes => {
+  if (bytes === 0) return '0';
+  const exp = Math.floor(Math.log(bytes) / Math.log(1000));
+  const size = bytes / 1000 ** exp;
+  const short = Math.round(size, 2);
+  const unit = SIZE_UNITS[exp];
+  return short + unit;
+};
+
+module.exports = { bytesToSize };
